moved stylesheets to new directory
