+++ src/components/App.js
@@ -5,7 +5,7 @@ import Library from './library';
 import Landing from './landing';
 import Recommendations from './recommendations';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import './App.css';
+import './styles/App.css';
 
 class App extends Component {
   render() {
