+++ src/components/footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './footer.css';
+import './styles/footer.css';
 
 export default function Footer() {
   return (
