+++ src/components/styles/game.css
