+++ src/components/styles/platform.css
