+++ src/components/styles/stats.css
